auth_service/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi-pagination==0.12.26
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
orjson==3.10.7
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.8
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
Werkzeug==3.0.4
